requirements.txt
requests-html==0.10.0
tqdm==4.67.1
pymongo==4.11.1
python-dotenv==1.0.1
lxml-html-clean==0.4.1